# docs/swagger.yaml
definitions:
...
  request.UpdateOrder:
    properties:
      address_id:
        type: integer
      delivery_status:
        type: string
      order_id:
        type: integer
      payment_status:
        type: string
      paymentmethod_id:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Glamgrove is an E-commerce platform for purchasing high quality beauty
    products
  title: Go + Gin E-Commerce API Glamgrove
  version: 1.0.0
paths:
...
  /admin/dashboard/salesReport:
    get:
      consumes:
      - application/json
      description: Generates a sales report based on the provided start and end dates,
      parameters:
      - description: Number of items per page
        in: query
        name: count
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_number
        required: true
        type: integer
      - description: Start date of the sales report (YYYY-MM-DD)
        in: query
        name: startDate
        required: true
        type: string
      - description: End date of the sales report (YYYY-MM-DD)
        in: query
        name: endDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated pdf
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Please add start date as params" or "Please add end date as
            params" or "There is no sales report on this period
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to generate PDF
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Generate sales report in PDF format
      tags:
      - Reports
...
  /admin/order/listOrder:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all orders with pagination
      parameters:
      - description: Number of orders per page
        in: query
        name: count
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get Orders successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Something went wrong!
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all orders
      tags:
      - Orders
...
  /order/cancelOrder:
    delete:
      consumes:
      - application/json
      description: Cancels an order with the specified order_id
      parameters:
      - description: ID of the order to be canceled
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted order
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Please add id as params" or "Can't delete order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Cancel an order
      tags:
      - Orders
  /order/createOrder:
    post:
      consumes:
      - application/json
      description: Creates an order with the provided parameters
      parameters:
      - description: ID of the address associated with the order
        in: query
        name: address_id
        required: true
        type: integer
      - description: ID of the payment method used for the order
        in: query
        name: paymentmethod_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created order. Please complete payment
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to get address id" or "Failed to get payment method
            id" or "Failed to get total amount" or "Failed to create order
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create an order
      tags:
      - Orders
  /order/listOrder:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all orders for the authenticated user with
        pagination
      parameters:
      - description: Number of orders per page
        in: query
        name: count
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get Orders successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Something went wrong!
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: List all orders for a user
      tags:
      - Orders
  /order/placeOrder:
    post:
      consumes:
      - application/json
      description: Places an order with the specified order_id and coupon_id
      parameters:
      - description: ID of the order to be placed
        in: query
        name: order_id
        required: true
        type: integer
      - description: ID of the coupon to be applied
        in: query
        name: coupon_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully placed order
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid coupon" or "Add more quantity" or "Failed to place
            order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Place an order
      tags:
      - Orders
  /order/updateOrder:
    put:
      consumes:
      - application/json
      description: Updates details of an existing order
      parameters:
      - description: Order details to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated order
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error while getting data from users" or "Error while updating
            data
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update an order
      tags:
      - Orders
...
  /return/product:
    post:
      consumes:
      - application/json
      description: Requests to return an order with the specified order ID, along
        with return reason (optional).
      parameters:
      - description: ID of the order to be returned
        in: query
        name: orderId
        required: true
        type: integer
      - description: Reason for return (optional)
        in: query
        name: Damage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully requested to return products
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Please add order id as params" or "Error while getting id from
            cookie" or "Invalid order_id" or "Failed to find refund amount" or "Failed
            to return order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Request to return an order
      tags:
      - Returns
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"